+++ Dockerfile-neuron
@@ -30,7 +30,7 @@ COPY --from=planner /usr/src/recipe.json recipe.json
 
 RUN --mount=type=secret,id=actions_results_url,env=ACTIONS_RESULTS_URL \
     --mount=type=secret,id=actions_runtime_token,env=ACTIONS_RUNTIME_TOKEN \
-    cargo chef cook --release --features python --no-default-features --recipe-path recipe.json && sccache -s
+    cargo chef cook --release --features python-neuron --no-default-features --recipe-path recipe.json && sccache -s
 
 COPY backends backends
 COPY core core
@@ -48,7 +48,7 @@ FROM builder AS http-builder
 
 RUN --mount=type=secret,id=actions_results_url,env=ACTIONS_RESULTS_URL \
     --mount=type=secret,id=actions_runtime_token,env=ACTIONS_RUNTIME_TOKEN \
-    cargo build --release --bin text-embeddings-router -F python -F http --no-default-features && sccache -s
+    cargo build --release --bin text-embeddings-router -F python-neuron -F http --no-default-features && sccache -s
 
 FROM builder AS grpc-builder
 
@@ -56,7 +56,7 @@ COPY proto proto
 
 RUN --mount=type=secret,id=actions_results_url,env=ACTIONS_RESULTS_URL \
     --mount=type=secret,id=actions_runtime_token,env=ACTIONS_RUNTIME_TOKEN \
-    cargo build --release --bin text-embeddings-router -F grpc -F python --no-default-features && sccache -s
+    cargo build --release --bin text-embeddings-router -F grpc -F python-neuron --no-default-features && sccache -s
 
 FROM public.ecr.aws/docker/library/ubuntu:22.04 AS neuron
