+++ router/Cargo.toml
@@ -86,6 +86,7 @@ metal = ["text-embeddings-backend/metal"]
 mkl = ["text-embeddings-backend/mkl", "dep:intel-mkl-src"]
 accelerate = ["text-embeddings-backend/accelerate"]
 python = ["text-embeddings-backend/python"]
+python-neuron = ["text-embeddings-backend/python-neuron"]
 ort = ["text-embeddings-backend/ort"]
 candle = ["text-embeddings-backend/candle"]
 candle-cuda = ["candle", "text-embeddings-backend/flash-attn", "dep:cudarc"]
